feat: implement shopping cart and item management with DTOs and domain mappers

+++ src/main/java/com/altice/domain/usecases/product/RemoveProduct.java
@@ -2,7 +2,10 @@ package com.altice.domain.usecases.product;
 
 import java.util.UUID;
 
+import com.altice.domain.enums.EnumErrorCode;
 import com.altice.domain.repositories.IProductRepository;
+import com.altice.domain.utils.StringUtils;
+import com.altice.domain.utils.exception.AlticeException;
 
 public class RemoveProduct {
 
@@ -13,6 +16,9 @@ public class RemoveProduct {
     }
 
     public void execute(String id) {
+        if (StringUtils.isNullOrEmpty(id)) {
+             throw new AlticeException(EnumErrorCode.REQUIRED_FIELD_FOR, "id", "remove product");
+        }
         productRepository.remove(UUID.fromString(id));
     }
 }
